anthropic-claude-api-learning/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
numpy>=1.26.0
scikit-learn>=1.4.0
tqdm>=4.66.0
httpx>=0.27.0
pydantic>=2.7.0
